+++ blitz-build/src/components/activityFeed/ActivityViewAll.js
@@ -1,12 +1,11 @@
 import React, { useEffect, useState, useContext } from "react";
 import styled from "styled-components";
 
-import NewTask from '../dashboard/NewTask'
+import ActivityFeedTask from '../dashboard/ActivityFeedTask'
 
 //context 
 import taskContext from '../../contexts/tasks/TaskContext'
 import {  Link } from "react-router-dom";
-
 
 
 
@@ -67,7 +66,7 @@ return (
            }
            else if (wasMadeToday(item.createdAt) || item.isComplete === true  ) {
              return (
-              <NewTask item={item} key={item.id} />
+              <ActivityFeedTask item={item} key={item.id} />
              )
            }
              })}
